refactor: separate proc and cmd in different files
